Move create_app() into separate file
